Add DI

+++ RaceLibrary/Converters/RaceConverter.cs
@@ -58,13 +58,13 @@ namespace RaceLibrary.Converters
             var lapElements = GetRequiredProperty(root, "Laps").EnumerateArray();
             foreach (var lapElement in lapElements)
             {
-                var lapNumber = short.Parse(GetRequiredProperty(lapElement, "number").GetString()!);
+                var lapNumber = int.Parse(GetRequiredProperty(lapElement, "number").GetString()!);
                 var timingsElements = GetRequiredProperty(lapElement, "Timings").EnumerateArray();
                 foreach (var timingsElement in timingsElements)
                 {
                     Lap lap = new(
                         GetRequiredProperty(timingsElement, "driverId").GetString()!,
-                        short.Parse(GetRequiredProperty(timingsElement, "position").GetString()!),
+                        int.Parse(GetRequiredProperty(timingsElement, "position").GetString()!),
                         lapNumber,
                         MapLapTime(timingsElement)
                     );
